requirements.txt
Django>=4.2,<5.3
Pillow
python-dotenv
scikit-learn
numpy
PyMuPDF
python-docx
nltk
google-genai
